refactor(pptx): extract chart.rs from lib.rs (verbatim)
Chart element parsers parse_legacy_chart (c: namespace) and parse_chartex (cx: namespace) plus PptxColorResolver + its ooxml_common::chart::ColorResolver impl (used only by parse_legacy_chart), now pub(crate). The general colour grammar (parse_color_node/parse_color_node_tint/PptxSchemeResolver) stays in lib.rs; chart.rs imports parse_color_node + attr from crate. child/attr/parse_color_node marked pub(crate) in lib.rs. lib.rs uses chart::*.
